fix: TypeError on GET /leave/holidays/upcoming with an explicit limit
LeaveController::upcomingHolidays() passed the validated limit straight
to HolidayService::upcoming(int $limit) without casting - Laravel's
'integer' validation rule checks the format, it doesn't cast the value,
so it stayed a string (query params always arrive as strings over
HTTP). With strict_types=1, that threw a TypeError the moment a real
client sent ?limit=N at all - found via the mobile app actually calling
this endpoint with an explicit limit, something the existing test never
exercised (it only ever hit the `?? 5` fallback, a literal int).

Fixed with the same explicit (int) cast AttendanceController::history()
already uses for its own limit param. New regression test passes
?limit=2 explicitly rather than relying on the default.

+++ app/Http/Controllers/Api/LeaveController.php
@@ -91,6 +91,6 @@ final class LeaveController extends Controller
             'limit' => ['sometimes', 'integer', 'min:1', 'max:20'],
         ]);
 
-        return response()->json(['data' => $service->upcoming($validated['limit'] ?? 5)]);
+        return response()->json(['data' => $service->upcoming((int) ($validated['limit'] ?? 5))]);
     }
 }
